refactor: 시간 fk id 를 가지고 있는 예약 조회 시 객체 반환을 boolean 반환으로 변경

+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -37,10 +37,9 @@ public class ReservationTimeService {
     }
 
     public void deleteById(long id) {
-        reservationRepository.findByReservationTimeId(id)
-            .ifPresent(reservation -> {
-                throw new IllegalArgumentException("해당 시간을 사용하는 예약이 존재합니다. timeId: " + id);
-            });
+        if (reservationRepository.existByReservationTimeId(id)) {
+            throw new IllegalArgumentException("해당 시간을 사용하는 예약이 존재합니다.");
+        }
         repository.deleteById(id);
     }
 }
